DABOM-237 fix: 프록시 내부 호출 문제 해결
