+++ src/main/java/com/project/domain/usagerecord/infra/sse/SsePublisher.java
@@ -1,15 +1,12 @@
 package com.project.domain.usagerecord.infra.sse;
 
 import java.time.LocalDateTime;
-import java.util.Optional;
 import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.atomic.AtomicReference;
 
 import org.springframework.scheduling.annotation.Async;
-import org.springframework.scheduling.annotation.Scheduled;
 import org.springframework.stereotype.Service;
 
-import com.project.domain.family.infra.cache.FamilyCacheRepository;
 import com.project.domain.usagerecord.dto.response.RealtimeTotalUsageResponse;
 import com.project.domain.usagerecord.dto.response.RealtimeUsageByMemberResponse;
 import com.project.global.event.dto.usage.UsageRealtimePayload;
@@ -23,7 +20,6 @@ import lombok.extern.slf4j.Slf4j;
 public class SsePublisher {
 
     private final EmitterRegistry emitterRegistry;
-    private final FamilyCacheRepository familyCacheRepository;
 
     private final ConcurrentHashMap<Long, AtomicReference<LocalDateTime>>
             lastTotalBytesTimeByFamily = new ConcurrentHashMap<>();
@@ -106,44 +102,5 @@ public class SsePublisher {
     private AtomicReference<LocalDateTime> memberTsRef(Long familyId) {
         return lastTotalMemberBytesTimeByFamily.computeIfAbsent(
                 familyId, id -> new AtomicReference<>(LocalDateTime.MIN));
-    }
-
-    @Scheduled(fixedDelay = 1000)
-    public void pollAndPushIfChanged() {
-        for (Long familyId : emitterRegistry.activeFamilyIds()) {
-
-            Optional<Long> latestOpt = familyCacheRepository.findFamilyRemainingBytes(familyId);
-            if (latestOpt.isEmpty()) {
-                lastSeen.remove(familyId);
-                continue;
-            }
-
-            long remainingBytes = latestOpt.get();
-            Long prev = lastSeen.putIfAbsent(familyId, remainingBytes);
-
-            if (prev == null || prev.longValue() != remainingBytes) {
-                lastSeen.put(familyId, remainingBytes);
-
-                // 임시로 고정
-                long totalLimitBytes = 20000;
-                long totalUsedBytes = totalLimitBytes - remainingBytes;
-
-                UsageRealtimePayload payload =
-                        new UsageRealtimePayload(
-                                familyId,
-                                1L,
-                                totalUsedBytes,
-                                totalLimitBytes,
-                                remainingBytes,
-                                30.0,
-                                null,
-                                null,
-                                null);
-
-                LocalDateTime now = LocalDateTime.now();
-                this.pushTotalUsageBytes(payload, now);
-                this.pushMemberUsageBytes(payload, now);
-            }
-        }
     }
 }
